+++ src/core/Value.cpp
@@ -55,11 +55,22 @@ bool Value::operator>=(const Value& other) const noexcept {
     return value_.value() >= other.value_.value();
 }
 
-std::ostream& operator<<(std::ostream& os, const Value& value) {
+Value Value::parse(std::string_view value) {
+    return Value(std::stod(std::string(value)));
+}
+
+std::string Value::str() const {
+    std::ostringstream os;
+
     os << std::setprecision(
               std::numeric_limits<numeric::Real::Storage>::max_digits10)
-       << value.real().value();
-    return os;
+       << value_.value();
+
+    return os.str();
+}
+
+std::ostream& operator<<(std::ostream& os, const Value& value) {
+    return os << value.str();
 }
 
 }  // namespace numathap::core