+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Improved the symbolic simplifier by recognizing the identity `x - x = 0`.
 - Added comprehensive Python type stubs (`__init__.pyi`) for the public API.
 - Implemented factorial function using ! symbol.
+- Implemented Constant Folder and inserted in differentiate pipeline.
+- Included x^1 to x in Simplifier.
 
 
 ## Release v0.1.4